Moved get weather function to weather.js Updated visuals on views
