+++ app/controllers/poi.js
@@ -2,27 +2,7 @@ const Walk = require("../models/poi");
 const User = require("../models/user");
 const Image = require("../models/image");
 const ImageStore = require("../utils/image-store");
-const axios = require("axios");
-
-async function getWeather(poi) {
-  const apiKey = process.env.api_key;
-  let weather = {};
-  const weatherRequest = `https://api.openweathermap.org/data/2.5/weather?lat=${poi.lat}&lon=${poi.lon}&appid=${apiKey}`;
-  const response = await axios.get(weatherRequest);
-  if (response.status == 200) {
-    weather = response.data;
-  }
-  const report = {
-    feelsLike: Math.round(weather.main.feels_like - 273.15),
-    clouds: weather.weather[0].description,
-    windSpeed: weather.wind.speed,
-    windDirection: weather.wind.deg,
-    visibility: weather.visibility / 1000,
-    humidity: weather.main.humidity,
-  };
-  //console.log("report", report);
-  return report;
-}
+const Weather = require("../utils/weather");
 
 const Poi = {
   home: {
@@ -78,10 +58,7 @@ const Poi = {
     handler: async function (request, h) {
       try {
         const poi = await Walk.findById({ _id: request.params._id }).populate("image").lean();
-        const apiKey = process.env.api_key;
-        console.log("api: ", apiKey);
-        console.log("poi: ", poi);
-        const weather = await getWeather(poi);
+        const weather = await Weather.getWeather(poi);
         console.log("Weather", weather);
         return h.view("poi", {
           title: "Viewing Point of Interest",
@@ -97,10 +74,11 @@ const Poi = {
   viewUserPoi: {
     handler: async function (request, h) {
       const poi = await Walk.findById({ _id: request.params._id }).populate("image").lean();
-      console.log("poi: ", poi);
+      const weather = await Weather.getWeather(poi);
       return h.view("user-poi", {
         title: "Viewing Your Point of Interest",
         poi: poi,
+        weather: weather,
       });
     },
   },
