+++ app/views/partials/admin-user-list.hbs
@@ -6,7 +6,7 @@
         <thead>
         <th>Name</th>
         <th>Email</th>
-        <th>Remove</th>
+        <th>Delete User</th>
         </thead>
         <tbody class="uk-text-left">
         {{#each user}}
