+++ app/views/partials/view-poi.hbs
@@ -18,7 +18,7 @@
             <td>{{poi.lat}}</td>
             <td>{{poi.lon}}</td>
             <td>{{poi.category}}</td>
-            <td>Feels Like: {{weather.feelsLike}}°C, Sky: {{weather.clouds}}</td>
+            <td>Feels Like: {{weather.feelsLike}}°C<br> Sky: {{weather.clouds}}</td>
         </tr>
         </tbody>
     </table>
