+++ app/views/update-poi.hbs
@@ -6,14 +6,11 @@
 <div class="uk-container uk-margin">
     <div class="uk-child-width-expand uk-flex-center uk-flex-middle uk-text-center" uk-grid>
         <div class="uk-width-auto@m">
-            <img width="300" src="/images/poi.png">
-        </div>
-        <div class="uk-width-expand@m">
             <div class="uk-card uk-card-default uk-width-xlarge uk-card-body uk-box-shadow-large">
                 <h3 class="uk-card-title uk-text-center">Poi Details: </h3>
                 {{> update-poi-form}}
             </div>
-            <div>
+            <div class="uk-card uk-card-default uk-width-xlarge uk-card-body uk-box-shadow-large">
                 {{> upload-image}}
             </div>
         </div>
